pyproject.toml
[tool.poetry]
name = "bsllc-discord-bot"
version = "0.1.0"
description = "One of the best Discord Bots, made by your's truely, BSLLC"
authors = ["ethanlchristensen <ethanlchristensen@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
"discord.py" = "^2.3.0"
python-dotenv = "^1.0.0"
python-weather = "^1.0.3"
bruhcolor = "^0.0.56"
matplotlib = "^3.7.1"
prettytable = "^3.8.0"
wcwidth = "^0.2.6"
PyNaCl = "^1.5.0"
cffi = "^1.15.1"
pycparser = "^2.21"
discordwebhook = "^1.0.3"
ruff = "^0.0.276"
pylint = "^2.17.4"
black = "^23.3.0"
invoke = "^2.1.3"
youtube-dl = "^2021.12.17"
flake8 = "^6.0.0"
mypy = "^1.4.1"
bandit = "^1.7.5"
gtts = "^2.3.2"
openai = "0.28"
bruhanimate = "0.1.92"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
